+++ src/main/java/christmas/utils/DateValidator.java
@@ -1,0 +1,21 @@
+package christmas.utils;
+
+public class DateValidator {
+    private static final int MIN_DAY = 1;
+    private static final int MAX_DAY = 31;
+    private static final String INVALID_DATE_MESSAGE = "[ERROR] 유효하지 않은 날짜입니다. 다시 입력해 주세요.";
+
+    public int validateDateNonNumeric(String input) {
+        try {
+            return Integer.parseInt(input);
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException(INVALID_DATE_MESSAGE);
+        }
+    }
+
+    public void validateDateRange(int day) {
+        if (day < MIN_DAY || day > MAX_DAY) {
+            throw new IllegalArgumentException(INVALID_DATE_MESSAGE);
+        }
+    }
+}
