+++ src/main/java/christmas/utils/OrderValidator.java
@@ -6,40 +6,19 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-public class InputValidator {
+public class OrderValidator {
     private static final int ZERO_ORDER_QUANTITY = 0;
-    private static final int MIN_DAY = 1;
-    private static final int MAX_DAY = 31;
     private static final int MIN_ORDER_QUANTITY = 1;
     private static final int MAX_ORDER_TOTAL_QUANTITY = 20;
-
-    private static final String menuPattern = "^[가-힣]+-[0-9]+$";
+    private static final String Pattern = "^[가-힣]+-[0-9]+$";
     private static final String UNKNOWN_CATEGORY = "NONE";
-    private static final String INVALID_DATE_MESSAGE = "[ERROR] 유효하지 않은 날짜입니다. 다시 입력해 주세요.";
     private static final String INVALID_ORDER_MESSAGE = "[ERROR] 유효하지 않은 주문입니다. 다시 입력해 주세요.";
-    private static final String ORDER_LIMIT_EXCEEDED_ERROR_MESSAGE = "[ERROR] 주문은 최대 20개까지 가능합니다. 다시 입력해 주세요.";
+    private static final String ORDER_LIMIT_EXCEEDED_MESSAGE = "[ERROR] 주문은 최대 20개까지 가능합니다. 다시 입력해 주세요.";
+
     private final Menu menu;
 
-    public InputValidator() {
+    public OrderValidator() {
         menu = new Menu();
-    }
-
-    public int validateNonNumeric(String input, String error) {
-        try {
-            return Integer.parseInt(input);
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException(error);
-        }
-    }
-
-    public int validateDateNonNumeric(String input) {
-        return validateNonNumeric(input, INVALID_DATE_MESSAGE);
-    }
-
-    public void validateDateRange(int day) {
-        if (day < MIN_DAY || day > MAX_DAY) {
-            throw new IllegalArgumentException(INVALID_DATE_MESSAGE);
-        }
     }
 
     public void validateMenuExistence(String food) {
@@ -55,11 +34,15 @@ public class InputValidator {
     }
 
     public int validateQuantityNonNumeric(String input) {
-        return validateNonNumeric(input, INVALID_ORDER_MESSAGE);
+        try {
+            return Integer.parseInt(input);
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException(INVALID_ORDER_MESSAGE);
+        }
     }
 
     public void validateMenuFormat(String food) {
-        if (!food.matches(menuPattern)) {
+        if (!food.matches(Pattern)) {
             throw new IllegalArgumentException(INVALID_ORDER_MESSAGE);
         }
     }
@@ -99,7 +82,7 @@ public class InputValidator {
                 .mapToInt(OrderMenuVO::quantity)
                 .sum();
         if (totalQuantity > MAX_ORDER_TOTAL_QUANTITY) {
-            throw new IllegalArgumentException(ORDER_LIMIT_EXCEEDED_ERROR_MESSAGE);
+            throw new IllegalArgumentException(ORDER_LIMIT_EXCEEDED_MESSAGE);
         }
     }
 }
